Start SpanContextPropagator refactor.
