+++ tracer/src/Datadog.Trace/ClrProfiler/ManualTracer.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using Datadog.Trace.DuckTyping;
 using Datadog.Trace.Logging;
+using Datadog.Trace.Propagators;
 using Datadog.Trace.Util;
 
 namespace Datadog.Trace.ClrProfiler
