+++ tracer/src/Datadog.Trace/Propagators/SpanContextPropagator.cs
@@ -15,7 +15,7 @@ using Datadog.Trace.Headers;
 using Datadog.Trace.Logging;
 using Datadog.Trace.Util;
 
-namespace Datadog.Trace
+namespace Datadog.Trace.Propagators
 {
     internal class SpanContextPropagator
     {
