+++ tracer/test/Datadog.Trace.Tests/SpanContextPropagatorTests.cs
@@ -7,6 +7,7 @@ using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Globalization;
 using Datadog.Trace.Headers;
+using Datadog.Trace.Propagators;
 using FluentAssertions;
 using Moq;
 using Xunit;
